# Conflicts:
#	bootcamp/activities/models.py
#	bootcamp/feeds/static/js/feeds.js
#	bootcamp/feeds/views.py

+++ bootcamp/feeds/views.py
@@ -133,16 +133,16 @@ def post(request):
 @ajax_required
 def like(request):
     feed_id = request.POST['feed']
-    feed = Feed.objects.get(pk=feed_id)
+    feed = Feed.objects.get(pk=feed_id).get_source_feed()
     user = request.user
-    like = Activity.objects.filter(activity_type=Activity.LIKE, feed=feed_id,
+    like = Activity.objects.filter(activity_type=Activity.LIKE, feed=feed.id,
                                    user=user)
     if like:
         user.profile.unotify_liked(feed)
         like.delete()
 
     else:
-        like = Activity(activity_type=Activity.LIKE, feed=feed_id, user=user)
+        like = Activity(activity_type=Activity.LIKE, feed=feed.id, user=user)
         like.save()
         user.profile.notify_liked(feed)
 
@@ -152,16 +152,16 @@ def like(request):
 @ajax_required
 def spam(request):
     feed_id = request.POST['feed']
-    feed = Feed.objects.get(pk=feed_id)
+    feed = Feed.objects.get(pk=feed_id).get_source_feed()
     user = request.user
-    spam = Activity.objects.filter(activity_type=Activity.SPAM, feed=feed_id,
+    spam = Activity.objects.filter(activity_type=Activity.SPAM, feed=feed.id,
                                    user=user)
     if spam:
         user.profile.unotify_spamed(feed)
         spam.delete()
 
     else:
-        spam = Activity(activity_type=Activity.SPAM, feed=feed_id, user=user)
+        spam = Activity(activity_type=Activity.SPAM, feed=feed.id, user=user)
         spam.save()
         user.profile.notify_spamed(feed)
 
@@ -172,7 +172,7 @@ def spam(request):
 def comment(request):
     if request.method == 'POST':
         feed_id = request.POST['feed']
-        feed = Feed.objects.get(pk=feed_id)
+        feed = Feed.objects.get(pk=feed_id).get_source_feed()
         post = request.POST['post']
         post = post.strip()
         if len(post) > 0:
@@ -187,7 +187,7 @@ def comment(request):
 
     else:
         feed_id = request.GET.get('feed')
-        feed = Feed.objects.get(pk=feed_id)
+        feed = Feed.objects.get(pk=feed_id).get_source_feed()
         return render(request, 'feeds/partial_feed_comments.html',
                       {'feed': feed})
 
@@ -205,6 +205,8 @@ def update(request):
     dump = {}
     for feed in feeds:
         dump[feed.pk] = {'likes': feed.likes, 'spams': feed.spams, 'comments': feed.comments}
+        dump[feed.pk] = {'likes': feed.get_likes_count(), 'comments': feed.get_comments_count()}
+
     data = json.dumps(dump)
     return HttpResponse(data, content_type='application/json')
 
@@ -213,7 +215,7 @@ def update(request):
 @ajax_required
 def track_comments(request):
     feed_id = request.GET.get('feed')
-    feed = Feed.objects.get(pk=feed_id)
+    feed = Feed.objects.get(pk=feed_id).get_source_feed()
     if len(feed.get_comments()) > 0:
         return render(
             request, 'feeds/partial_feed_comments.html', {'feed': feed})
@@ -229,13 +231,14 @@ def remove(request):
         feed_id = request.POST.get('feed')
         feed = Feed.objects.get(pk=feed_id)
         if feed.user == request.user:
-            likes = feed.get_likes()
-            spams = feed.get_spams()
             parent = feed.parent
-            for like in likes:
-                like.delete()
-            for spam in spams:
-                spam.delete()
+            if not feed.is_retweet():
+                likes = feed.get_likes()
+                spams = feed.get_spams()
+                for like in likes:
+                    like.delete()
+                for spam in spams:
+                    spam.delete()
             feed.delete()
             if parent:
                 parent.calculate_comments()
@@ -247,3 +250,22 @@ def remove(request):
 
     except Exception:
         return HttpResponseBadRequest()
+
+@login_required
+@ajax_required
+def retweet(request):
+    last_feed = request.POST.get('last_feed')
+    source_feed_id = request.POST.get('feed')
+    csrf_token = (csrf(request)['csrf_token'])
+    user = request.user
+
+    source_feed = Feed.objects.get(pk=source_feed_id)
+    source_feed.retweet(user)
+
+    html = _html_feeds(last_feed, user, csrf_token)
+    return HttpResponse(html)
+
+@login_required
+@ajax_required
+def remove_retweet(request):
+    return remove(request)
